- Good compile on tclserver changes for online monitoring. - Fix some 64 bit unclean sizing problems. - Update bootstrap to work with autoreconf.

+++ llnlReadout/devices/C830.cpp
@@ -260,10 +260,10 @@ C830::Initialize(CVMUSB& controller)
   // Fire off the list.. it's all writes so the buffer is just a dummy long.
   
 
-  uint32_t dummyLong;
+  size_t dummyLong;
 
   controller.executeList(initList,
-			 &dummyLong, sizeof(uint32_t), &dummyLong);
+			 &dummyLong, sizeof(size_t), &dummyLong);
 
 
 }
